Removed unused LineProtocalSyntax class
